Correcting value conversion error that prevent splash potions from rolling back correctly. bug 242

+++ src/me/botsko/prism/actions/ItemStackAction.java
@@ -113,7 +113,7 @@ public class ItemStackAction extends GenericAction {
 			if (!TypeUtils.isNumeric(blockArr[0])) return;
 			int block_id = Integer.parseInt(blockArr[0]);
 			if (blockArr.length >= 3){
-				int block_subid = (byte) Integer.parseInt(blockArr[1]);
+				int block_subid = Integer.parseInt(blockArr[1]);
 				quantity = Integer.parseInt(blockArr[2]);
 				item = new ItemStack(block_id,quantity,(short)block_subid);
 				if(blockArr.length > 3){
